Split LogProfile into LogProfile and LogOutput classes
