+++ config/http-logger.php
@@ -3,10 +3,16 @@
 return [
 
     /*
-     * The log profile used to log requests. A log profile determines wheter a request will be,
-     * logged or not, and how the message is formatted. It should implement `LogProfile``.
+     * The log profile which determines whether a request should be logged.
+     * It should implement `LogProfile`.
      */
     'log_profile' => \Spatie\HttpLogger\DefaultLogProfile::class,
+
+    /**
+     * The log profile used to log the actual request.
+     * It should implement `LogOutput`.
+     */
+    'log_output' => \Spatie\HttpLogger\DefaultLogOutput::class,
 
     /*
      * Filter out body fields which will never be logged.
