+++ src/LogOutput.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Spatie\HttpLogger;
+
+use Illuminate\Http\Request;
+
+interface LogOutput
+{
+    public function logRequest(Request $request);
+}
