+++ src/LogProfile.php
@@ -7,6 +7,4 @@ use Illuminate\Http\Request;
 interface LogProfile
 {
     public function shouldLogRequest(Request $request): bool;
-
-    public function logRequest(Request $request);
 }
