+++ src/Middlewares/HttpLogger.php
@@ -4,21 +4,24 @@ namespace Spatie\HttpLogger\Middlewares;
 
 use Closure;
 use Illuminate\Http\Request;
+use Spatie\HttpLogger\LogOutput;
 use Spatie\HttpLogger\LogProfile;
 
 class HttpLogger
 {
     protected $logProfile;
+    protected $logOutput;
 
-    public function __construct(LogProfile $logProfile)
+    public function __construct(LogProfile $logProfile, LogOutput $logOutput)
     {
         $this->logProfile = $logProfile;
+        $this->logOutput = $logOutput;
     }
 
     public function handle(Request $request, Closure $next)
     {
         if ($this->logProfile->shouldLogRequest($request)) {
-            $this->logProfile->logRequest($request);
+            $this->logOutput->logRequest($request);
         }
 
         return $next($request);
